Fix email preprocessing in publish endpoint
Fixed a bug in the publish endpoint where the application crashes when the CURATOR_MAIL environment variable is not set.
This was done by modifying the logic on how we validate and use the CURATOR_MAIL value, allowing the system to work even when the environment variable is not present.
Also, added a log message to trace the recipient of the email generated.